backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Data Validation and Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and ORM
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Vector Database
qdrant-client==2.7.0

# LLM Integration
google-generativeai==0.3.1

# Environment Management
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.12.1
pylint==3.0.3
mypy==1.7.1

# Utilities
python-jose==3.3.0
cryptography==41.0.7
